Support multimodal model & fix llama_ppl dtype err & fix llama quantized slower err

+++ lightllm/models/llama_ppl/layer_infer/transformer_layer_infer.py
@@ -40,7 +40,7 @@ class LlamaPPlTransformerLayerInfer(LlamaTransformerLayerInfer):
                                                           infer_state.mem_manager.key_scale_buffer[self.layer_num_],
                                                           infer_state.mem_manager.value_buffer[self.layer_num_],
                                                           infer_state.mem_manager.value_scale_buffer[self.layer_num_],
-                                                          infer_state.b_loc,
+                                                          infer_state.b_loc.to(torch.int64),
                                                           infer_state.b_seq_len,
                                                           infer_state.max_len_in_batch)
